Fixes the setup test

+++ tests/test_pulseaudio_setup_wrapper.sh
@@ -17,8 +17,10 @@
 
 STREAM="stream1"
 SINK="audiosync"
+CURDIR=$(dirname "$0")
 
-# First of all, starting pulseaudio if necessary
+# First of all, restarting pulseaudio
+killall pulseaudio
 if ! pulseaudio --check &>/dev/null; then
     pulseaudio --start -D
 fi
@@ -29,17 +31,51 @@ pacmd load-module module-loopback source="dummy.monitor" sink_input_properties="
 num_sinks=$(pactl list short sinks | wc -l)
 
 # Runinng the pulseaudio setup
-./tests/test_pulseaudio_setup "$STREAM"
+echo "==== First run ===="
+"$CURDIR/test_pulseaudio_setup" "$STREAM"
 
 # Only one new sink should be created
-[ "$((num_sinks + 1))" -eq $(pactl list short sinks | wc -l) ] || echo "Sink creation unsuccessful"; exit 1
+if [ "$((num_sinks + 1))" -ne $(pactl list short sinks | wc -l) ]; then
+    echo "Sink creation unsuccessful"
+    exit 1
+fi
+
+# Checking that the sink was created
+if ! pacmd list-sinks | grep "<$SINK>" -B 3 -q; then
+    echo "Sink not found"
+    exit 1
+fi
+
+# Making sure that the input was correctly moved
+if ! pacmd list-sink-inputs | grep "$STREAM" -B 20 | grep "sink:" | grep "<$SINK>" -q; then
+    echo "Stream not moved to the required sink"
+    exit 1
+fi
 
 # A second test to make sure that no new sinks are created, meaning that
 # the previous one is being re-used.
-[ "$num_sinks" -eq $(pactl list short sinks | wc -l) ] || echo "Sink not being reused"; exit 1
-./tests/test_pulseaudio_setup "$STREAM"
+echo "==== Second run ===="
+"$CURDIR/test_pulseaudio_setup" "$STREAM"
 
-# Restarting pulseaudio
+# Same test, making sure that no new sinks were created.
+if [ "$((num_sinks + 1))" -ne $(pactl list short sinks | wc -l) ]; then
+    echo "Sink not being reused"
+    exit 1
+fi
+
+# Checking again
+if ! pacmd list-sinks | grep "<$SINK>" -B 3 -q; then
+    echo "Sink not found"
+    exit 1
+fi
+
+# Making sure that the input was correctly moved again
+if ! pacmd list-sink-inputs | grep "$STREAM" -B 20 | grep "sink:" | grep "<$SINK>" -q; then
+    echo "Stream not moved to the required sink"
+    exit 1
+fi
+
+# Restarting pulseaudio at the end to undo the changes made.
 killall pulseaudio
 pulseaudio --start -D
 exit 0
